+++ test/regress.c
@@ -182,7 +182,7 @@ multiple_write_cb(evutil_socket_t fd, short event, void *arg)
 	int len;
 
 	len = 128;
-	if (woff + len >= sizeof(wbuf))
+	if (woff + len >= (int)sizeof(wbuf))
 		len = sizeof(wbuf) - woff;
 
 	len = write(fd, wbuf + woff, len);
@@ -195,7 +195,7 @@ multiple_write_cb(evutil_socket_t fd, short event, void *arg)
 
 	woff += len;
 
-	if (woff >= sizeof(wbuf)) {
+	if (woff >= (int)sizeof(wbuf)) {
 		shutdown(fd, SHUT_WR);
 		if (usepersist)
 			event_del(ev);
@@ -489,7 +489,7 @@ test_multiple(void)
 	/* Multiple read and write test */
 	setup_test("Multiple read/write: ");
 	memset(rbuf, 0, sizeof(rbuf));
-	for (i = 0; i < sizeof(wbuf); i++)
+	for (i = 0; i < (int)sizeof(wbuf); i++)
 		wbuf[i] = i;
 
 	roff = woff = 0;
@@ -518,7 +518,7 @@ test_persistent(void)
 	/* Multiple read and write test with persist */
 	setup_test("Persist read/write: ");
 	memset(rbuf, 0, sizeof(rbuf));
-	for (i = 0; i < sizeof(wbuf); i++)
+	for (i = 0; i < (int)sizeof(wbuf); i++)
 		wbuf[i] = i;
 
 	roff = woff = 0;
@@ -1681,7 +1681,7 @@ evtag_fuzz(void *ptr)
 	evtag_init();
 
 	for (j = 0; j < 100; j++) {
-		for (i = 0; i < sizeof(buffer); i++)
+		for (i = 0; i < (int)sizeof(buffer); i++)
 			buffer[i] = rand();
 		evbuffer_drain(tmp, -1);
 		evbuffer_add(tmp, buffer, sizeof(buffer));
